+++ src/main/java/com/iexec/blockchain/chain/QueueService.java
@@ -1,4 +1,20 @@
-package com.iexec.blockchain.tool;
+/*
+ * Copyright 2021-2024 IEXEC BLOCKCHAIN TECH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.iexec.blockchain.chain;
 
 import lombok.EqualsAndHashCode;
 import org.springframework.beans.factory.annotation.Value;
@@ -85,6 +101,7 @@ public class QueueService {
     @EqualsAndHashCode(callSuper = true)
     static class TaskWithPriority<T> extends FutureTask<T> implements Comparable<TaskWithPriority<?>> {
         private final BlockchainAction action;
+
         TaskWithPriority(Runnable task) {
             super(task, null);
             this.action = (BlockchainAction) task;