+++ src/main/java/com/iexec/blockchain/command/task/finalize/TaskFinalizeService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2024 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,13 +16,12 @@
 
 package com.iexec.blockchain.command.task.finalize;
 
-
+import com.iexec.blockchain.chain.QueueService;
 import com.iexec.blockchain.command.generic.CommandEngine;
-import com.iexec.blockchain.tool.QueueService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import static com.iexec.blockchain.tool.IexecHubService.isByte32;
+import static com.iexec.blockchain.chain.IexecHubService.isByte32;
 
 @Slf4j
 @Service
@@ -42,9 +41,8 @@ public class TaskFinalizeService extends CommandEngine<TaskFinalize, TaskFinaliz
             log.error("At least one bad args [chainTaskId:{}, args:{}]", chainTaskId, args);
             return "";
         }
-        return startBlockchainCommand(new TaskFinalizeArgs(chainTaskId,
-                args.getResultLink(),
-                args.getCallbackData()),
+        return startBlockchainCommand(
+                new TaskFinalizeArgs(chainTaskId, args.getResultLink(), args.getCallbackData()),
                 true);
     }