+++ src/test/java/com/iexec/blockchain/chain/QueueServiceTests.java
@@ -1,4 +1,20 @@
-package com.iexec.blockchain.tool;
+/*
+ * Copyright 2021-2024 IEXEC BLOCKCHAIN TECH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.iexec.blockchain.chain;
 
 import lombok.extern.slf4j.Slf4j;
 import org.awaitility.Awaitility;
@@ -16,8 +32,8 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
 @Slf4j
 class QueueServiceTests {
@@ -165,10 +181,14 @@ class QueueServiceTests {
     // endregion
 
     //region BlockchainAction
+    private void blockchainAction() {
+        // a blockchain action
+    }
+
     @Test
     void compareBlockchainActionAgainstNull() {
-        QueueService.BlockchainAction lowPriorityAction = new QueueService.BlockchainAction(() -> {}, false);
-        QueueService.BlockchainAction highPriorityAction = new QueueService.BlockchainAction(() -> {}, true);
+        QueueService.BlockchainAction lowPriorityAction = new QueueService.BlockchainAction(this::blockchainAction, false);
+        QueueService.BlockchainAction highPriorityAction = new QueueService.BlockchainAction(this::blockchainAction, true);
         assertThatThrownBy(() -> lowPriorityAction.compareTo(null))
                 .isInstanceOf(NullPointerException.class);
         assertThatThrownBy(() -> highPriorityAction.compareTo(null))
@@ -179,10 +199,10 @@ class QueueServiceTests {
     void validateBlockchainActionComparisons() {
         final int hasHigherPriority = -1;
         final int hasLowerPriority = 1;
-        QueueService.BlockchainAction action1 = new QueueService.BlockchainAction(() -> {}, false);
-        QueueService.BlockchainAction action2 = new QueueService.BlockchainAction(() -> {}, false);
-        QueueService.BlockchainAction action3 = new QueueService.BlockchainAction(() -> {}, true);
-        QueueService.BlockchainAction action4 = new QueueService.BlockchainAction(() -> {}, true);
+        QueueService.BlockchainAction action1 = new QueueService.BlockchainAction(this::blockchainAction, false);
+        QueueService.BlockchainAction action2 = new QueueService.BlockchainAction(this::blockchainAction, false);
+        QueueService.BlockchainAction action3 = new QueueService.BlockchainAction(this::blockchainAction, true);
+        QueueService.BlockchainAction action4 = new QueueService.BlockchainAction(this::blockchainAction, true);
         //check action1
         assertThat(action1).isEqualByComparingTo(action1);
         assertThat(action1.compareTo(action2)).isEqualTo(hasHigherPriority);
@@ -208,21 +228,23 @@ class QueueServiceTests {
 
     //region TaskWithPriority
     @Test
-    void compareTaskWithPriorityAgainstNul() {
-        QueueService.BlockchainAction lowPriorityAction = new QueueService.BlockchainAction(() -> {}, false);
-        QueueService.BlockchainAction highPriorityAction = new QueueService.BlockchainAction(() -> {}, true);
-        assertThatThrownBy(() -> lowPriorityAction.compareTo(null))
+    void compareTaskWithPriorityAgainstNull() {
+        QueueService.BlockchainAction lowPriorityAction = new QueueService.BlockchainAction(this::blockchainAction, false);
+        QueueService.BlockchainAction highPriorityAction = new QueueService.BlockchainAction(this::blockchainAction, true);
+        QueueService.TaskWithPriority<Runnable> lowPriorityTask = new QueueService.TaskWithPriority<>(lowPriorityAction);
+        QueueService.TaskWithPriority<Runnable> highPriorityTask = new QueueService.TaskWithPriority<>(highPriorityAction);
+        assertThatThrownBy(() -> lowPriorityTask.compareTo(null))
                 .isInstanceOf(NullPointerException.class);
-        assertThatThrownBy(() -> highPriorityAction.compareTo(null))
+        assertThatThrownBy(() -> highPriorityTask.compareTo(null))
                 .isInstanceOf(NullPointerException.class);
     }
 
     @Test
     void validateTaskWithPriorityComparisons() {
-        QueueService.BlockchainAction lowPriorityAction = new QueueService.BlockchainAction(() -> {}, false);
+        QueueService.BlockchainAction lowPriorityAction = new QueueService.BlockchainAction(this::blockchainAction, false);
         QueueService.TaskWithPriority<Runnable> task1 = new QueueService.TaskWithPriority<>(lowPriorityAction);
         QueueService.TaskWithPriority<Runnable> task2 = new QueueService.TaskWithPriority<>(lowPriorityAction);
-        assertThat(task1.compareTo(task2)).isZero();
+        assertThat(task1).isEqualByComparingTo(task2);
     }
     //endregion
 }