+++ src/test/java/com/iexec/blockchain/chain/WalletConfigurationTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.iexec.blockchain.tool;
+package com.iexec.blockchain.chain;
 
 import com.iexec.commons.poco.chain.SignerService;
 import org.junit.jupiter.api.Test;